+++ RELEASENOTES.md
@@ -1,6 +1,8 @@
 These are the release notes for the [mediawiki-datamodel](README.md).
 
 ## Version 0.5 (under development)
+* Implemented Log class
+* Implemented LogList class
 
 
 ## Version 0.4 (2014-07-08)